# Backend/openapi.yml
openapi: 3.0.0
info:
  title: Beacon API
  version: 1.0.0
  description: API for managing pins in the Beacon application
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/trip/plan:
    post:
      summary: Plan a trip
      description: Generate a sustainable trip plan with transit options and AI itinerary
      operationId: planTrip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - startLocation
                - endLocation
                - itineraryType
              properties:
                startLocation:
                  type: string
                  description: Starting city
                  example: "San Francisco"
                endLocation:
                  type: string
                  description: Destination city
                  example: "Los Angeles"
                itineraryType:
                  type: string
                  description: Type of itinerary
                  example: "nature"
                departureDate:
                  type: string
                  description: Departure date (YYYY-MM-DD)
                  example: "2026-02-01"
                durationDays:
                  type: integer
                  description: Trip duration in days
                  example: 3
      responses:
        200:
          description: Trip plan generated successfully
          content:
            application/json:
              schema:
                type: object
        400:
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "creatorID is required"
        details:
          type: string
          description: Additional error details
          example: "FOREIGN KEY constraint failed"
...